ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.3
groq==0.11.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2